- added edit and delete icons to all admin panels - added confirm messages to delete

+++ src/app/component/admin/teams/teams.component.ts
@@ -23,6 +23,7 @@ export class TeamsComponent implements OnInit {
   }
 
   delete(id: string) {
+    if (!confirm(`Are you sure you want to delete this match? You won't be able to restore it.`)) return;
     this.teamsService.deleteTeam(id);
   }
 }
